Predict University admission using logistic regression
Suppose you are the administrator of a university department and you want to
determine the probability of admission of each candidate based on their two-way results
exams. - You have historical data from previous candidates that you
can use as a training set for logistic regression. - For each training example, you have the scores of the two candidates
examinations and the admission decision. - Your task is to build a classification model that estimates the probability
admission of a candidate based on the scores obtained in these two exams.